[Arch] Moved decompress pointers to their respective block file
